fix(observability): windows build + distinguish tenant-pool removal reasons
- process metrics: move RLIMIT_NOFILE lookup behind build tags
  (process_unix.go / process_other.go). syscall.Getrlimit/RLIMIT_NOFILE are
  undefined on Windows, so the untagged file broke `GOOS=windows` builds even
  though process metrics are Linux-only at runtime.
- tenant pool: thread a removal reason through removeLocked so the metric no
  longer counts replacement/invalidation/shutdown as cache evictions. Emits
  drive9_service_operations_total{component="tenant_pool",operation="remove",
  result="evict|replace|invalidate|shutdown"}; only result="evict" reflects
  real max-size cache pressure.
